refactor(approach): split banner + manifest under the 400-line cap (#2382)
#2382 pushed two files over the #1680 400-line cap and grew a
grandfathered file:
- trip_recording_banner.dart (410→351): extracted the PiP palette
  resolution (BannerPalette + approachOverlayPalette + bandPalette) to
  trip_recording_banner_palette.dart — a distinct, context-free concern.
- feature_manifest.dart (416→350): extracted FeatureManifestEntry to its
  own file (re-exported, so existing importers are unaffected).
- feature_management_section.dart snapshot 706→711: +5 for the three
  per-feature switch cases of Feature.approachOverlay (intrinsic
  per-feature growth; documented in the file_length snapshot).

Fixes the l10n-gate + test(3)/(1) file_length failures on this PR.
